this is for implemtation of the change of frame feature
I am using the function QGraphicsView::centerOn(qreal x, qreal y) to change the view coordinated as soon as each fram is finished

WorldScene::worldDataChanged() was modified( modelIndex of the currently selected item is taken and its coordinates are
passed to the QGraphicsView::centerOn() function

At the end of each frame as the position of the WorldGraphicsItems are updated the positin of the view also
changes to the current location of the currently selected item of the model.

I tried to add an option "change frame" in the View Menu of the Step MainWindow
that option is conected to A protected slot in the WorldScene i.e. toggleFrameChange() which toggles the bool variable
doFrameChange (protected in the WorldScene)

The view opsition is updated in the WorldScene::worldDataChanged() only when the bool variable doFrameChange is set to true

However there seems to be a littel problem with this arrangement at the moment so it has been commented.

To experience this feature. place two diks on the scene, set different velocities to both of them. press the simulate button
now if you click on any disk you will experience the relative motion between the two moving disks

the enable-disable of frame-change feature will soon be enforced

Signed-off-by: Aniket Anvit <seeanvit@gmail.com>