A dialog added to ask the user to input the values of the friction and restitution coefficients
Here's how to see what has been implemented in this branch:

Create two bodies on the scene ( preferably take a box) anchor it , put weightforce on the scene and create another body (preferably a disk) .
Using control key, select both the bodies . the go to Edit->Friction
In the dialog that appears enter the values of the coefficients.

no you can see the effect of varying the coefficient of restitution between two bodies.

This way any pair of bodies can have different value of the bounceness/restitution or friction coefficient
Signed-off-by: Aniket Anvit <seeanvit@gmail.com>